# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "AutomationStatus": {
        "properties": {
          "candidate_count": {
            "default": 0,
            "title": "Candidate Count",
            "type": "integer"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "enabled_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled At"
          },
          "interval_minutes": {
            "default": 15,
            "title": "Interval Minutes",
            "type": "integer"
          },
          "last_run_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run At"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "AutomationStatus",
        "type": "object"
      },
      "AutomationUpdate": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "interval_minutes": {
            "anyOf": [
              {
                "maximum": 1440.0,
                "minimum": 5.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Minutes"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "AutomationUpdate",
        "type": "object"
      },
...
      "BackupConfigurationUpdate": {
        "properties": {
          "components": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Components",
            "type": "object"
          }
        },
        "title": "BackupConfigurationUpdate",
        "type": "object"
      },
      "BackupCreateRequest": {
        "properties": {
          "components": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "boolean"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Components"
          }
        },
        "title": "BackupCreateRequest",
        "type": "object"
      },
      "BackupRestoreRequest": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "BackupRestoreRequest",
        "type": "object"
      },
...
      "DanbooruCredentialsUpdate": {
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username"
        ],
        "title": "DanbooruCredentialsUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Keivotos - Danbooru",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/automation": {
      "get": {
        "operationId": "get_automation_api_automation_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Automation"
      },
      "put": {
        "operationId": "update_automation_api_automation_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutomationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Automation"
      }
    },
    "/api/backups": {
      "get": {
        "operationId": "get_backup_configuration_api_backups_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Backup Configuration"
      },
      "put": {
        "operationId": "configure_backups_api_backups_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupConfigurationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Configure Backups"
      }
    },
    "/api/backups/create": {
      "post": {
        "operationId": "create_metadata_backup_api_backups_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Metadata Backup"
      }
    },
    "/api/backups/estimate": {
      "post": {
        "operationId": "estimate_backup_api_backups_estimate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Estimate Backup"
      }
    },
    "/api/backups/restore": {
      "post": {
        "operationId": "restore_metadata_backup_api_backups_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupRestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Metadata Backup"
      }
    },
    "/api/backups/{backup_name}/inspect": {
      "get": {
        "operationId": "inspect_metadata_backup_api_backups__backup_name__inspect_get",
        "parameters": [
          {
            "in": "path",
            "name": "backup_name",
            "required": true,
            "schema": {
              "title": "Backup Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inspect Metadata Backup"
      }
    },
...
    "/api/danbooru/credentials": {
      "delete": {
        "operationId": "delete_danbooru_credentials_api_danbooru_credentials_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DanbooruCredentialStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Delete Danbooru Credentials"
      },
      "get": {
        "operationId": "get_danbooru_credentials_api_danbooru_credentials_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DanbooruCredentialStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Danbooru Credentials"
      },
      "put": {
        "operationId": "update_danbooru_credentials_api_danbooru_credentials_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DanbooruCredentialsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DanbooruCredentialStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Danbooru Credentials"
      }
    },
...
    "/api/import-pipeline/run": {
      "post": {
        "operationId": "run_import_api_import_pipeline_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Import"
      }
    },
    "/api/import-pipeline/task": {
      "get": {
        "operationId": "import_pipeline_task_api_import_pipeline_task_get",
        "parameters": [
          {
            "in": "query",
            "name": "after_index",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Index"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolStatusInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Pipeline Task"
      }
    },
...
    "/api/thumbnails/cache/limit": {
      "put": {
        "operationId": "update_thumbnail_limit_api_thumbnails_cache_limit_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThumbnailCacheLimitUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Thumbnail Limit"
      }
    },
...
    "/api/tools/backfill/run": {
      "post": {
        "operationId": "run_backfill_tool_api_tools_backfill_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackfillToolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolRunResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Backfill Tool"
      }
    },
...
    "/api/tools/{tool_id}/cancel": {
      "post": {
        "operationId": "cancel_tool_api_tools__tool_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "tool_id",
            "required": true,
            "schema": {
              "title": "Tool Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolRunResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Tool"
      }
    },
    "/api/tools/{tool_id}/run": {
      "post": {
        "operationId": "run_tool_api_tools__tool_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "tool_id",
            "required": true,
            "schema": {
              "title": "Tool Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolRunResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Tool"
      }
    },
    "/api/tools/{tool_id}/status": {
      "get": {
        "operationId": "tool_status_api_tools__tool_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "tool_id",
            "required": true,
            "schema": {
              "title": "Tool Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolStatusInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tool Status"
      }
    },
...
  }
}